add partial order flow

+++ application/controllers/CheckOut/CheckoutController.php
@@ -66,41 +66,72 @@ class CheckoutController extends CI_Controller
     public function placeOrder()
     {
 
+        $debug = array();
 
         $cart_contents = $this->cart->contents();
-
         $total_amt = $this->cart->total(); //Temperory set to cart total
         $this->data['zones'] = $this->setting->getByKey('shipping_settings')['value'];
         $weights = $this->totalWeights();
-
-        $order_data = array(
-            'user_id' => /*($this->session->userdata('user')->id) ? $this->session->userdata('user')->id :~*/
-                $this->data['user']->id,//Logged in user id
-            'delivery_address' => $this->input->post('delivery_address'),
-            'pincode' => $this->input->post('pincode'),
-            'order_status' => '1',
-            'total_amt' => $total_amt
-        );
-        $order_id = $this->OrdersModel->insert($order_data);
-
-
-        foreach ($cart_contents as $product) {
-            $order_product_data = array(
-                'product_id' => $product['id'],
-                'order_id' => $order_id,
-                'qty' => $product['qty'],
+        if ($this->input->server('REQUEST_METHOD') == 'POST') {
+            $this->form_validation->set_rules(
+                'deliveryId',
+                'Delivery Address',
+                'required|is_natural_no_zero',
+                array('is_natural_no_zero' => 'An address is required to checkout')
             );
-            $this->OrderProductModel->insert($order_product_data);
-        }
-        if ($order_id) {
-            $this->cart->destroy();
-            $this->session->set_flashdata('success', 'Order placed successfully');
-            redirect(base_url('index.php/shop'));
-            exit;
-        } else {
-            $this->session->set_flashdata('error', 'Oops! Something went wrong.');
-            redirect(base_url('index.php/shop'));
-            exit;
+            $inputs = $this->input->post();
+
+            array_push($debug, array('method' => 'post', 'inputs' => $inputs));
+
+            if ($this->form_validation->run()) {
+
+                array_push($debug, array('formvalidationrun' => 'true'));
+
+                $order_data = array(
+                    'user_id' => $this->data['user']->id,
+                    'deliveryId' => $this->input->post('deliveryId'),
+                    'order_status' => 'ordered',
+                    'total_amt' => $total_amt,
+                    'total_weight' => $weights->total
+                );
+                $order_id = $this->OrdersModel->insert($order_data);
+                $index = 0;
+
+                foreach ($cart_contents as $key => $product) {
+
+                    $order_product_data = array(
+                        'product_id' => $product['id'],
+                        'order_id' => $order_id,
+                        'qty' => $product['qty'],
+                        'weight' => $weights->totalWeights[$index] * $product['qty']
+                    );
+                    $index++;
+
+                    $this->OrderProductModel->insert($order_product_data);
+                }
+                if ($order_id) {
+                    $this->cart->destroy();
+                    $this->session->set_flashdata('success', 'Order placed successfully');
+                    redirect(base_url('shop/orders'));
+                    exit;
+                } else {
+                    $this->session->set_flashdata('error', 'Oops! Something went wrong.');
+                    redirect(base_url('shop/orders'));
+                    exit;
+                }
+            } else {
+                $this->data['addresses'] = $this->AddressModel->getAddressByUserId($this->data['user']->user_id);
+                $this->data['debug'] = $debug;
+                $weights = $this->totalWeights();
+                $this->data['content'] = 'checkout/checkout';
+                $this->data['zones'] = $this->setting->getByKey('shipping_settings')['value'];
+                $this->data['weights'] = $weights->totalWeights;
+                $this->data['totalWeight'] = $weights->total;
+
+                $this->load->templateProfile('checkout/checkout', $this->data);
+            }
+
+
         }
